+++ lib/karafka/web/ui/pro/app.rb
@@ -103,12 +103,16 @@ module Karafka
             r.on 'errors' do
               controller = Controllers::Errors.new(params)
 
-              r.get Integer do |partition_id|
-                render_response controller.index(partition_id)
-              end
-
               r.get Integer, Integer do |partition_id, offset|
                 render_response controller.show(partition_id, offset)
+              end
+
+              r.get Integer do |partition_id|
+                render_response controller.partition(partition_id)
+              end
+
+              r.get do
+                render_response controller.index
               end
             end