+++ lib/karafka/web/ui/pro/controllers/consumers.rb
@@ -22,7 +22,7 @@ module Karafka
             def index
               @current_state = Models::State.current!
               @counters = Models::Counters.new(@current_state)
-              @processes, last_page = Lib::PaginateArray.call(
+              @processes, last_page = Lib::Paginations::Paginators::Arrays.call(
                 Models::Processes.active(@current_state),
                 @params.current_page
               )