+++ lib/karafka/web/ui/pro/controllers/errors.rb
@@ -18,12 +18,42 @@ module Karafka
         module Controllers
           # Errors details controller
           class Errors < Ui::Controllers::Base
+            def index
+              @topic_id = errors_topic
+              @partitions_count = Models::ClusterInfo.partitions_count(errors_topic)
+              @max_aggregable_partitions = Web.config.ui.explorer.max_aggregable_partitions
+
+              # For topics with a lot of partitions we cannot get all the data efficiently, that
+              # is why we limit number of partitions by default
+              if @partitions_count > @max_aggregable_partitions
+                aggreagable_partitions = 0...@max_aggregable_partitions
+                @limited = true
+              else
+                aggreagable_partitions = 0...@partitions_count
+                @limited = false
+              end
+
+              @error_messages, next_page = Models::Message.topic_page(
+                errors_topic, aggreagable_partitions, @params.current_page
+              )
+
+              paginate(@params.current_page, next_page)
+
+              respond
+            end
+
             # @param partition_id [Integer] id of the partition of errors we are interested in
-            def index(partition_id)
+            def partition(partition_id)
               @partition_id = partition_id
               @watermark_offsets = Ui::Models::WatermarkOffsets.find(errors_topic, @partition_id)
+              @partitions_count = Models::ClusterInfo.partitions_count(errors_topic)
 
-              previous_offset, @error_messages, next_offset, @partitions_count = current_page_data
+              previous_offset, @error_messages, next_offset = Models::Message.offset_page(
+                errors_topic,
+                @partition_id,
+                @params.current_offset,
+                @watermark_offsets
+              )
 
               paginate(
                 previous_offset,
@@ -52,18 +82,6 @@ module Karafka
             end
 
             private
-
-            # @return [Array] Array with requested messages as well as pagination details and other
-            #   obtained metadata
-            def current_page_data
-              Models::Message.offset_page(
-                errors_topic,
-                @partition_id,
-                @params.current_offset,
-                @watermark_offsets[:low],
-                @watermark_offsets[:high]
-              )
-            end
 
             # @return [String] errors topic
             def errors_topic